0.2.0 in progress
add more docs
add partial eq & eq for structs
