+++ src/lib.rs
@@ -4,11 +4,11 @@
 //!
 //! Current capability:
 //!  - Reading & parsing STDF files.
+//!  - Reading & parsing ATDF files.
 //!  - Support gzip and bzip compressed files.
 //!
 //! In development:
 //!  - (dev) Dump `StdfRecord` to a new stdf file.
-//!  - (dev) Reading & parsing ATDF file (Text format).
 //!  - (dev) Functions for ATDF <-> STDF format.
 
 // lib.rs
@@ -44,248 +44,7 @@ pub mod atdf_file;
 mod tests {
     use crate::*;
     use atdf_types::atdf_record_field::*;
-    use stdf_record_type::*;
     use stdf_types::ByteOrder;
-
-    #[test]
-    fn record_default_value_test() {
-        let empty_raw_data = [0u8; 0];
-        // mir
-        let mir_rec = StdfRecord::new(REC_MIR);
-        if let StdfRecord::MIR(ref inner) = mir_rec {
-            assert_eq!(inner.mode_cod, ' ', "Testing default of new rec");
-            assert_eq!(inner.rtst_cod, ' ', "Testing default of new rec");
-            assert_eq!(inner.prot_cod, ' ', "Testing default of new rec");
-            assert_eq!(inner.burn_tim, 65535, "Testing default of new rec");
-            assert_eq!(inner.cmod_cod, ' ', "Testing default of new rec");
-        }
-        let mir_rec =
-            mir_rec.read_from_bytes(&empty_raw_data, &stdf_types::ByteOrder::LittleEndian);
-        if let StdfRecord::MIR(ref inner) = mir_rec {
-            assert_eq!(inner.mode_cod, ' ', "Testing default value after reading");
-            assert_eq!(inner.rtst_cod, ' ', "Testing default value after reading");
-            assert_eq!(inner.prot_cod, ' ', "Testing default value after reading");
-            assert_eq!(inner.burn_tim, 65535, "Testing default value after reading");
-            assert_eq!(inner.cmod_cod, ' ', "Testing default value after reading");
-        }
-
-        // mrr
-        let mrr_rec = StdfRecord::new(REC_MRR);
-        if let StdfRecord::MRR(ref inner) = mrr_rec {
-            assert_eq!(inner.disp_cod, ' ', "Testing default of new rec");
-        }
-        let mrr_rec =
-            mrr_rec.read_from_bytes(&empty_raw_data, &stdf_types::ByteOrder::LittleEndian);
-        if let StdfRecord::MRR(ref inner) = mrr_rec {
-            assert_eq!(inner.disp_cod, ' ', "Testing default value after reading");
-        }
-
-        // pcr
-        let pcr_rec = StdfRecord::new(REC_PCR);
-        if let StdfRecord::PCR(ref inner) = pcr_rec {
-            assert_eq!(inner.rtst_cnt, 4_294_967_295, "Testing default of new rec");
-            assert_eq!(inner.abrt_cnt, 4_294_967_295, "Testing default of new rec");
-            assert_eq!(inner.good_cnt, 4_294_967_295, "Testing default of new rec");
-            assert_eq!(inner.func_cnt, 4_294_967_295, "Testing default of new rec");
-        }
-        let pcr_rec =
-            pcr_rec.read_from_bytes(&empty_raw_data, &stdf_types::ByteOrder::LittleEndian);
-        if let StdfRecord::PCR(ref inner) = pcr_rec {
-            assert_eq!(
-                inner.rtst_cnt, 4_294_967_295,
-                "Testing default value after reading"
-            );
-            assert_eq!(
-                inner.abrt_cnt, 4_294_967_295,
-                "Testing default value after reading"
-            );
-            assert_eq!(
-                inner.good_cnt, 4_294_967_295,
-                "Testing default value after reading"
-            );
-            assert_eq!(
-                inner.func_cnt, 4_294_967_295,
-                "Testing default value after reading"
-            );
-        }
-
-        // hbr
-        let hbr_rec = StdfRecord::new(REC_HBR);
-        if let StdfRecord::HBR(ref inner) = hbr_rec {
-            assert_eq!(inner.hbin_pf, ' ', "Testing default of new rec");
-        }
-        let hbr_rec =
-            hbr_rec.read_from_bytes(&empty_raw_data, &stdf_types::ByteOrder::LittleEndian);
-        if let StdfRecord::HBR(ref inner) = hbr_rec {
-            assert_eq!(inner.hbin_pf, ' ', "Testing default value after reading");
-        }
-
-        // sbr
-        let sbr_rec = StdfRecord::new(REC_SBR);
-        if let StdfRecord::SBR(ref inner) = sbr_rec {
-            assert_eq!(inner.sbin_pf, ' ', "Testing default of new rec");
-        }
-        let sbr_rec =
-            sbr_rec.read_from_bytes(&empty_raw_data, &stdf_types::ByteOrder::LittleEndian);
-        if let StdfRecord::SBR(ref inner) = sbr_rec {
-            assert_eq!(inner.sbin_pf, ' ', "Testing default value after reading");
-        }
-
-        // pmr
-        let pmr_rec = StdfRecord::new(REC_PMR);
-        if let StdfRecord::PMR(ref inner) = pmr_rec {
-            assert_eq!(inner.chan_typ, 0, "Testing default of new rec");
-            assert_eq!(inner.head_num, 1, "Testing default of new rec");
-            assert_eq!(inner.site_num, 1, "Testing default of new rec");
-        }
-        let pmr_rec =
-            pmr_rec.read_from_bytes(&empty_raw_data, &stdf_types::ByteOrder::LittleEndian);
-        if let StdfRecord::PMR(ref inner) = pmr_rec {
-            assert_eq!(inner.chan_typ, 0, "Testing default value after reading");
-            assert_eq!(inner.head_num, 1, "Testing default value after reading");
-            assert_eq!(inner.site_num, 1, "Testing default value after reading");
-        }
-
-        // cdr
-        let cdr_rec = StdfRecord::new(REC_CDR);
-        if let StdfRecord::CDR(ref inner) = cdr_rec {
-            assert_eq!(inner.inv_val, 255, "Testing default of new rec");
-        }
-        let cdr_rec =
-            cdr_rec.read_from_bytes(&empty_raw_data, &stdf_types::ByteOrder::LittleEndian);
-        if let StdfRecord::CDR(ref inner) = cdr_rec {
-            assert_eq!(inner.inv_val, 255, "Testing default value after reading");
-        }
-
-        // wir
-        let wir_rec = StdfRecord::new(REC_WIR);
-        if let StdfRecord::WIR(ref inner) = wir_rec {
-            assert_eq!(inner.site_grp, 255, "Testing default of new rec");
-        }
-        let wir_rec =
-            wir_rec.read_from_bytes(&empty_raw_data, &stdf_types::ByteOrder::LittleEndian);
-        if let StdfRecord::WIR(ref inner) = wir_rec {
-            assert_eq!(inner.site_grp, 255, "Testing default value after reading");
-        }
-
-        // wrr
-        let wrr_rec = StdfRecord::new(REC_WRR);
-        if let StdfRecord::WRR(ref inner) = wrr_rec {
-            assert_eq!(inner.site_grp, 255, "Testing default of new rec");
-            assert_eq!(inner.rtst_cnt, 4_294_967_295, "Testing default of new rec");
-            assert_eq!(inner.abrt_cnt, 4_294_967_295, "Testing default of new rec");
-            assert_eq!(inner.good_cnt, 4_294_967_295, "Testing default of new rec");
-            assert_eq!(inner.func_cnt, 4_294_967_295, "Testing default of new rec");
-        }
-        let wrr_rec =
-            wrr_rec.read_from_bytes(&empty_raw_data, &stdf_types::ByteOrder::LittleEndian);
-        if let StdfRecord::WRR(ref inner) = wrr_rec {
-            assert_eq!(inner.site_grp, 255, "Testing default value after reading");
-            assert_eq!(
-                inner.rtst_cnt, 4_294_967_295,
-                "Testing default value after reading"
-            );
-            assert_eq!(
-                inner.abrt_cnt, 4_294_967_295,
-                "Testing default value after reading"
-            );
-            assert_eq!(
-                inner.good_cnt, 4_294_967_295,
-                "Testing default value after reading"
-            );
-            assert_eq!(
-                inner.func_cnt, 4_294_967_295,
-                "Testing default value after reading"
-            );
-        }
-
-        // wcr
-        let wcr_rec = StdfRecord::new(REC_WCR);
-        if let StdfRecord::WCR(ref inner) = wcr_rec {
-            assert_eq!(inner.wafr_siz, 0.0, "Testing default of new rec");
-            assert_eq!(inner.die_ht, 0.0, "Testing default of new rec");
-            assert_eq!(inner.die_wid, 0.0, "Testing default of new rec");
-            assert_eq!(inner.wf_units, 0, "Testing default of new rec");
-            assert_eq!(inner.wf_flat, ' ', "Testing default of new rec");
-            assert_eq!(inner.center_x, -32768, "Testing default of new rec");
-            assert_eq!(inner.center_y, -32768, "Testing default of new rec");
-            assert_eq!(inner.pos_x, ' ', "Testing default of new rec");
-            assert_eq!(inner.pos_y, ' ', "Testing default of new rec");
-        }
-        let wcr_rec =
-            wcr_rec.read_from_bytes(&empty_raw_data, &stdf_types::ByteOrder::LittleEndian);
-        if let StdfRecord::WCR(ref inner) = wcr_rec {
-            assert_eq!(inner.wafr_siz, 0.0, "Testing default value after reading");
-            assert_eq!(inner.die_ht, 0.0, "Testing default value after reading");
-            assert_eq!(inner.die_wid, 0.0, "Testing default value after reading");
-            assert_eq!(inner.wf_units, 0, "Testing default value after reading");
-            assert_eq!(inner.wf_flat, ' ', "Testing default value after reading");
-            assert_eq!(
-                inner.center_x, -32768,
-                "Testing default value after reading"
-            );
-            assert_eq!(
-                inner.center_y, -32768,
-                "Testing default value after reading"
-            );
-            assert_eq!(inner.pos_x, ' ', "Testing default value after reading");
-            assert_eq!(inner.pos_y, ' ', "Testing default value after reading");
-        }
-
-        // prr
-        let prr_rec = StdfRecord::new(REC_PRR);
-        if let StdfRecord::PRR(ref inner) = prr_rec {
-            assert_eq!(inner.soft_bin, 65535, "Testing default of new rec");
-            assert_eq!(inner.x_coord, -32768, "Testing default of new rec");
-            assert_eq!(inner.y_coord, -32768, "Testing default of new rec");
-            assert_eq!(inner.test_t, 0, "Testing default of new rec");
-        }
-        let prr_rec =
-            prr_rec.read_from_bytes(&empty_raw_data, &stdf_types::ByteOrder::LittleEndian);
-        if let StdfRecord::PRR(ref inner) = prr_rec {
-            assert_eq!(inner.soft_bin, 65535, "Testing default value after reading");
-            assert_eq!(inner.x_coord, -32768, "Testing default value after reading");
-            assert_eq!(inner.y_coord, -32768, "Testing default value after reading");
-            assert_eq!(inner.test_t, 0, "Testing default value after reading");
-        }
-
-        // tsr
-        let tsr_rec = StdfRecord::new(REC_TSR);
-        if let StdfRecord::TSR(ref inner) = tsr_rec {
-            assert_eq!(inner.test_typ, ' ', "Testing default of new rec");
-            assert_eq!(inner.exec_cnt, 4_294_967_295, "Testing default of new rec");
-            assert_eq!(inner.fail_cnt, 4_294_967_295, "Testing default of new rec");
-            assert_eq!(inner.alrm_cnt, 4_294_967_295, "Testing default of new rec");
-        }
-        let tsr_rec =
-            tsr_rec.read_from_bytes(&empty_raw_data, &stdf_types::ByteOrder::LittleEndian);
-        if let StdfRecord::TSR(ref inner) = tsr_rec {
-            assert_eq!(inner.test_typ, ' ', "Testing default value after reading");
-            assert_eq!(
-                inner.exec_cnt, 4_294_967_295,
-                "Testing default value after reading"
-            );
-            assert_eq!(
-                inner.fail_cnt, 4_294_967_295,
-                "Testing default value after reading"
-            );
-            assert_eq!(
-                inner.alrm_cnt, 4_294_967_295,
-                "Testing default value after reading"
-            );
-        }
-
-        // ftr
-        let ftr_rec = StdfRecord::new(REC_FTR);
-        if let StdfRecord::FTR(ref inner) = ftr_rec {
-            assert_eq!(inner.patg_num, 255, "Testing default of new rec");
-        }
-        let ftr_rec =
-            ftr_rec.read_from_bytes(&empty_raw_data, &stdf_types::ByteOrder::LittleEndian);
-        if let StdfRecord::FTR(ref inner) = ftr_rec {
-            assert_eq!(inner.patg_num, 255, "Testing default value after reading");
-        }
-    }
 
     // unsigned data type
     #[test]
